Storage rework Part 2

+++ liberty/handlers/audio.py
@@ -3,6 +3,7 @@ import asyncio
 import shutil
 from pathlib import Path
 import logging
+import os
 
 from utils.downloader import YouTubeVideo, download
 from utils.constants import _AUDIO_DIR, _DEEP_FRIED_FFMPEG_OPTIONS
@@ -41,11 +42,12 @@ class AudioHandler:
 
 
     async def download_and_add_song(self, url, channel_id=None, manipulation=None, added_by=None):
-        video = download(url)
+        video = await download(url)
+        logger.info(f'Successfully downloaded {url}')
         return await self.add_song(video.path, channel_id, manipulation, video.title, source='Downloaded Youtube Audio', added_by=added_by)
 
 
-    async def save_song(self, name, queue_position=None):
+    async def save_song(self, name, queue_position=None, bucket_handler=None):
         if not self.playing:
             logger.warning('Cannot save song if nothing is playing')
             return False
@@ -56,7 +58,15 @@ class AudioHandler:
             song = self.queue._queue[queue_position]
         old_path = song.path
         new_path = _AUDIO_DIR + name
+        if os.path.exists(new_path):
+            logger.warning(f'Not saving song file {new_path} because a file already exists there')
+            return False
         shutil.copyfile(old_path, new_path)
+        # Upload saved song to the cloud for persistence if we have buckets
+        if bucket_handler:
+            logger.info(f'Uploading {name} to cloud storage')
+            resp = await bucket_handler.upload_file(new_path, 'audio/' + name)
+        logger.info(f'Successfully saved {name}')
         return True
